modifying the initialization. modifying some keywords for parameters. adding a descrete mapping from asphericity for aspect ratio for spheroids. ready to be used for spheroids like general case just by specifying the asphericity.

+++ main.cc
@@ -17,7 +17,7 @@ void Initialize(){
 
 void Run(){
 
-	CSys sys(config.get_param<size_t>("nParticle"));
+	CSys sys(config.get_param<unsigned int>("nParticle"));
 	sys.initialize(config);
 	sys.solve();
 	}
